fix(dir): Rename path to dir
