+++ crates/snapbox/src/data/source.rs
@@ -63,7 +63,7 @@ impl From<Inline> for DataSource {
 impl std::fmt::Display for DataSource {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match &self.inner {
-            DataSourceInner::Path(value) => crate::path::display_relpath(value).fmt(f),
+            DataSourceInner::Path(value) => crate::dir::display_relpath(value).fmt(f),
             DataSourceInner::Inline(value) => value.fmt(f),
         }
     }
