+++ crates/snapbox/src/dir/ops.rs
